requirements.txt
fastapi==0.74.1
paho-mqtt==1.5.1
uvicorn[standard]==0.18.0
